+++ app/helpers/exercise_evaluations_table_helper.rb
@@ -160,4 +160,31 @@ module ExerciseEvaluationsTableHelper
       link_to 'Create', '#', class: "tiny button exercise-evaluations-table-create-submission", data: {student_group_id: student_group.id, url: exercise_evaluation_path(@exercise), confirm: confirm_message}
     end
   end
+
+  def evaluation_input_field(f)
+    rating = f.object.rating
+
+    f.input_field(:value, evaluation_input_field_options(rating))
+  end
+
+  def evaluation_input_field_options(rating)
+    options = {}
+
+    if rating.is_a? BinaryRating
+      options[:as] = :boolean
+      options[:data] = {customforms: 'disabled'}
+      options[:class] = 'no-margin'
+    elsif rating.is_a? ValueNumberRating
+      options[:type] = :number
+      options[:step] = 1
+      options[:min] = rating.min_value
+      options[:max] = rating.max_value
+    elsif rating.is_a? ValuePercentRating
+      options[:type] = :number
+      options[:step] = 5
+      options[:min] = rating.min_value
+      options[:max] = rating.max_value
+    end
+    options
+  end
 end
