--- app/policies/submission_evaluation_policy.rb
@@ -1,31 +1,0 @@
-class SubmissionEvaluationPolicy < PunditBasePolicy
-  def show?
-    user.admin? ||
-    user.lecturer_of_term?(record.submission.exercise.term) ||
-    user.tutor_of_tutorial_group?(record.submission.student_group_registration.student_group.tutorial_group)
-  end
-
-  def new?
-    user.admin? ||
-    user.lecturer_of_term?(record.submission.exercise.term) ||
-    user.tutor_of_tutorial_group?(record.submission.student_group_registration.student_group.tutorial_group)
-  end
-
-  def create?
-    user.admin? ||
-    user.lecturer_of_term?(record.submission.exercise.term) ||
-    user.tutor_of_tutorial_group?(record.submission.student_group_registration.student_group.tutorial_group)
-  end
-
-  def edit?
-    user.admin? ||
-    user.lecturer_of_term?(record.submission.exercise.term) ||
-    user.tutor_of_tutorial_group?(record.submission.student_group_registration.student_group.tutorial_group)
-  end
-
-  def update?
-    user.admin? ||
-    user.lecturer_of_term?(record.submission.exercise.term) ||
-    user.tutor_of_tutorial_group?(record.submission.student_group_registration.student_group.tutorial_group)
-  end
-end
