--- app/views/submission_evaluations/_form.html.erb
@@ -1,16 +1,0 @@
-<%= simple_form_for @submission_evaluation, url: submission_evaluation_path(current_submission, @submission_evaluation), html: {class: "custom"} do |se| %>
-  <% prepared_evaluations.each do |rating_group, evaluations| %>
-    <h3><%= rating_group.title %> (<%= pluralize rating_group.points, "point"%>)</h3>
-    <% evaluations.each do |evaluation| %>
-      <p>
-        <%= se.fields_for :evaluations, evaluation do |f| %>
-        <%= f.association :rating, as: :hidden %>
-        <%= f.input :type, as: :hidden %>
-        <%= evaluation_input_field(f) %>
-        <% end %>
-      </p>
-    <% end %>
-  <% end %>
-  
-  <%= se.submit class: "button" %>
-<% end %>
