--- app/views/submission_evaluations/_sidebar.html.erb
@@ -1,3 +1,0 @@
-<h4>Submitted Files</h4>
-
---- show submitted files here ---
