--- app/views/submission_evaluations/edit.html.erb
@@ -1,7 +1,0 @@
-<% title "Update Evaluation for #{current_submission.exercise.title} of #{current_submission.student.fullname}" %>
-
-<% sidebar do %>
-  <%= render "sidebar" %>
-<% end %>
-
-<%= render "form" %>
