--- app/views/submission_evaluations/new.html.erb
@@ -1,6 +1,0 @@
-<% title "New Evaluation for #{current_submission.exercise.title} of #{current_submission.student.fullname}" %>
-<% sidebar do %>
-  <%= render "sidebar" %>
-<% end %>
-
-<%= render "form" %>
